zng_app_context

Struct ContextLocal

pub struct ContextLocal<T: Send + Sync + 'static> { /* private fields */ }

Represents an AppLocal<T> value that can be temporarily overridden in a context.

The context works across threads, as long as the threads are instrumented using LocalContext.

Use the context_local! macro to declare a static variable in the same style as thread_local!.

Implementations

impl<T: Send + Sync + 'static> ContextLocal<T>

pub fn id(&'static self) -> AppLocalId

Gets an ID for this context local instance that is valid for the lifetime of the process.

Note that comparing two &'static CTX_LOCAL pointers is incorrect, because in "hot_reload" builds the statics can be different and still represent the same app local. This ID identifies the actual inner pointer.

pub fn with_context<R>( &'static self, value: &mut Option<Arc<T>>, f: impl FnOnce() -> R, ) -> R

Calls f with the value loaded in context.

The value is moved into context, f is called, then the value is moved back to value.

Panics

Panics if value is None.

pub fn with_context_var<R>( &'static self, value: &mut Option<Arc<T>>, f: impl FnOnce() -> R, ) -> R

Same as with_context, but value represents a variable.

Values loaded with this method are captured by CaptureFilter::ContextVars.

pub fn with_default<R>(&'static self, f: impl FnOnce() -> R) -> R

Calls f with no value loaded in context.

pub fn is_default(&'static self) -> bool

Gets if no value is set in the context.

pub fn get(&'static self) -> Arc<T>

Clone a reference to the current value in the context or the default value.

pub fn get_clone(&'static self) -> T
where T: Clone,

Clone the current value in the context or the default value.

impl<T: Send + Sync + 'static> ContextLocal<RwLock<T>>

pub fn read_only(&'static self) -> ReadOnlyRwLock<T>

Gets a read-only shared reference to the current context value.

pub fn read(&'static self) -> RwLockReadGuardOwned<T>

Locks this RwLock with shared read access, blocking the current thread until it can be acquired.

See parking_lot::RwLock::read for more details.

pub fn read_recursive(&'static self) -> RwLockReadGuardOwned<T>

Locks this RwLock with shared read access, blocking the current thread until it can be acquired.

Unlike read, this method is guaranteed to succeed without blocking if another read lock is held at the time of the call.

See parking_lot::RwLock::read for more details.

pub fn write(&'static self) -> RwLockWriteGuardOwned<T>

Locks this RwLock with exclusive write access, blocking the current thread until it can be acquired.

See parking_lot::RwLock::write for more details.

pub fn try_read(&'static self) -> Option<RwLockReadGuardOwned<T>>

Try lock this RwLock with shared read access, blocking the current thread until it can be acquired.

See parking_lot::RwLock::try_read for more details.

pub fn try_read_recursive(&'static self) -> Option<RwLockReadGuardOwned<T>>

Locks this RwLock with shared read access, blocking the current thread until it can be acquired.

See parking_lot::RwLock::try_read_recursive for more details.

pub fn try_write(&'static self) -> Option<RwLockWriteGuardOwned<T>>

Locks this RwLock with exclusive write access, blocking the current thread until it can be acquired.

See parking_lot::RwLock::try_write for more details.