zng_app_context
Trait ContextLocalKeyProvider

pub trait ContextLocalKeyProvider {
    // Required method
    fn context_local_key(&'static self) -> AppLocalId;
}

Provides an identifying key for a context local value.

Implemented by all ContextLocal<T> already, only implement this for context local thin wrappers.

Required Methods

fn context_local_key(&'static self) -> AppLocalId

Gets the key.