Crate zng_app_context

This crate is part of the zng project.

App execution context.

Macros

app_local

Declares new app local variable.

context_local

Declares new app and context local variable.

Structs

AppId

Identifies an app instance.

AppLocal

An app local storage.

AppScope

Represents an app lifetime, ends the app on drop.

ContextLocal

Represents an AppLocal<T> value that can be temporarily overridden in a context.

ContextValueSet

Identifies a selection of LocalContext values.

FullLocalContext

A local context that can be loaded cheaply.

LocalContext

Tracks the current execution context.

MappedRwLockReadGuardOwned

Represents a read guard for an Arc<RwLock<T>> that owns a reference to it, mapped from another read guard.

MappedRwLockWriteGuardOwned

Represents a write guard for an Arc<RwLock<T>> that owns a reference to it, mapped from another read guard.

ReadOnlyRwLock

Read-only wrapper on an Arc<RwLock<T>> contextual value.

RunOnDrop

Helper, runs a cleanup action once on drop.

RwLockReadGuardOwned

Represents a read guard for an Arc<RwLock<T>> that owns a reference to it.

RwLockWriteGuardOwned

Represents a read guard for an Arc<RwLock<T>> that owns a reference to it.

StaticAppId

Lazy inited AppId.

TracingDispatcherContext

Represents the tracing::dispatcher::get_default dispatcher in a context value set.

Enums

CaptureFilter

Defines a LocalContext::capture_filtered filter.

Traits

ContextLocalKeyProvider

Provides an identifying key for a context local value.