Trait zng_app_context::ContextLocalKeyProvider

pub trait ContextLocalKeyProvider {
    // Required method
    fn context_local_key(&'static self) -> TypeId;
}

Provides an identifying key for a context local value.

Implemented by all ContextLocal<T> already, only implement this for context local thin wrappers.

Required Methods

fn context_local_key(&'static self) -> TypeId

Gets the key.